Integrated the RESTful module into the main server. I've also separated out the server code from `main.py` into its own file in `server.py`. This will allow the next steps of unittests for the server to be easier as the server module can be imported into the rest script.
To allow the RESTful async operations to work alongside the main server async operations, the server _run() function is now ran through the asyncio API run_until_complete. Previous _run() was being ran as the sole event loop on the thread so the RESTful event loop was not getting any processing time.
